+++ backend/src/main/java/ch/lonelyporter/statsapp/web/model/Statistic.java
@@ -23,7 +23,13 @@ public class Statistic {
     @NoArgsConstructor
     public static class Dataset {
         private String label;
+        private int index;
         private List<DataPoint> dataPoints = new ArrayList<>();
+
+        public Dataset(String label, List<DataPoint> dataPoints) {
+            this.label = label;
+            this.dataPoints = dataPoints;
+        }
     }
 
     @Getter